# Replace cudf's concurrent_ordered_map with cuco::static_map in semi/anti joins