--- a/cpp/src/join/semi_join.cu
+++ b/cpp/src/join/semi_join.cu
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-#include <hash/concurrent_unordered_map.cuh>
 #include <join/join_common_utils.cuh>
 #include <join/join_common_utils.hpp>