@vyasr 2021-11-23
> This PR depends on [NVIDIA/cuCollections#113](https://github.com/NVIDIA/cuCollections/pull/113), otherwise the default hash allocator won't work here.

I left a comment to this effect [here](https://github.com/rapidsai/cudf/pull/9666/files#diff-835047ebb75e5848893ff55834f3b834e7482fa258574ff839a305d882d217b9R111), but I agree that we probably shouldn't merge this until that PR is done to avoid unexpected races.