@PointKernel 2021-12-16
@vyasr Back to the benchmark results, any idea why the new implementation is slower?